refactor: move ListBox2D and ListBoxItem2D into containers.py per reviewer request
