+++ fury/ui/__init__.pyi
@@ -45,11 +45,10 @@ __all__ = [
 # from .containers import GridUI, ImageContainer2D, Panel2D, TabPanel2D, TabUI
 # from .core import UI, Button2D, Disk2D, Rectangle2D, TextBlock2D
 
-from .containers import Panel2D
+from .containers import ListBox2D, ListBoxItem2D, Panel2D
 from .context import UIContext
 from .core import UI, Anchor, Disk2D, Rectangle2D, TextBlock2D
 from .elements import LineSlider2D, TextButton2D, TexturedButton2D
-from .listbox2d import ListBox2D, ListBoxItem2D
 
 # from .elements import (
 #     Card2D,
