+++ fury/ui/tests/test_containers.py
@@ -5,6 +5,7 @@ from os.path import join as pjoin
 from PIL import Image
 import numpy as np
 import numpy.testing as npt
+import pytest
 
 from fury import ui, window
 
@@ -664,3 +665,54 @@ def test_panel2d_visual_snapshot(
 #     npt.assert_equal(0, tab_ui_bottom.active_tab_idx)
 #     npt.assert_equal(14, next(changes))
 #     npt.assert_equal(5, next(collapses))
+
+
+@pytest.mark.filterwarnings("ignore::RuntimeWarning")
+def test_listbox_creation():
+    lb = ui.ListBox2D(values=["a", "b", "c"])
+    assert lb.values == ["a", "b", "c"]
+    assert lb.selected == []
+
+
+@pytest.mark.filterwarnings("ignore::RuntimeWarning")
+def test_single_selection():
+    lb = ui.ListBox2D(values=["a", "b", "c"])
+    item = lb.slots[0]
+    item.element = "a"
+    lb.select(item)
+    assert lb.selected == ["a"]
+
+
+@pytest.mark.filterwarnings("ignore::RuntimeWarning")
+def test_multiselect():
+    lb = ui.ListBox2D(values=["a", "b", "c"], multiselection=True)
+    item1 = lb.slots[0]
+    item1.element = "a"
+    lb.select(item1, multiselect=True)
+
+    item2 = lb.slots[1]
+    item2.element = "b"
+    lb.select(item2, multiselect=True)
+
+    assert set(lb.selected) == {"a", "b"}
+
+
+@pytest.mark.filterwarnings("ignore::RuntimeWarning")
+def test_range_select():
+    lb = ui.ListBox2D(values=["a", "b", "c", "d"], multiselection=True)
+
+    item1 = lb.slots[0]
+    item1.element = "a"
+    lb.select(item1)
+
+    item2 = lb.slots[2]
+    item2.element = "c"
+    lb.select(item2, range_select=True)
+
+    assert lb.selected == ["a", "b", "c"]
+
+
+@pytest.mark.filterwarnings("ignore::RuntimeWarning")
+def test_scrollbar_hidden_when_not_needed():
+    lb = ui.ListBox2D(values=["a"])
+    assert lb.nb_slots >= len(lb.values)
